Every function in its own file, part 1 [#CLICKHOUSE-2]
